fix checkout of remote branches
now just checks out the local branch name instead which in modern Git actually creates a new local tracking branch beforehand named after the remote one if it didn't exist.

this was achieved by changing all branch actions from branch id (e.g. origin/master) to branch name (e.g. master). This commit also adds this new `id` field for all refs, and it changes the way the branches are determined: By asking git for it with `git branch -A`, but this shouldn't change any existing logic. Except now we can be sure if a branch is remote or not. This is reflected in the new field `branch.remote_name`.
Also refactors the tag name stuff: Now tags also have an id and a name field.