fix pull/push on local branches, even if they don't have a remote configured yet
the most likely remote will be prefilled: remote name (if branch itself is remote), tracking remote name (if branch itself is local but has set a remote) or default remote (the remote/tracking remote of master, main or the first branch)

in short, it should now always work as you'd expect it to. In standard projects, you'll see `origin` as the param for push/pull everywhere now and MISSING_REMOTE_NAME should be a very rare case.