finalize file diff tree view
in style and functionality and fixing bugs
dev: changes `stateful_computed` to return the config value as a promise instead, so the default will now only be used if nothing is stored in config. this makes sense because we don't need the default if the user had  already changed it to something else before, also the loading time is minimal. But this requires the creation of some components to be delayed, so the respective <Suspense> were added.
Also changed the default <details><summary> list icon
Adds a new external dependency, vueuse, to enable reusable templates