refactor message passing logic of persisted state: globalState, workspaceState, git state (repos, index)
now there's a single `state` object in `extension.coffee` backend which can handle all current and future sorts of persisted state. It comes at the expense of slightly less great type support, but prevents the need for a lot of duplication everywhere.
The web's counterpart of `state` is `stateful_computed`.