Add new default commit action: Edit message
Works also for older commits. Only works when your work tree is clean (no changed files). Only works when the selected commit is part of the current branch.

Unfortunately, prefilling the message field with the existing commit message would be too complex, so for now to *alter* something, you have to copy paste the existing one.
This is quite an unusual feature for any Git client to have and rewriting history should usually be avoided, so having this feature in the first place is rather rare already.

closes #87