tests with delegating parsing to extension backend
somehow the parsing is a bit slow here (not tested much), but the problem with this approach is the passing in post_message. both the implicit and the explicit json stringify take about as long as the parsing itself, and often fail with large data with "invalid string length". the imported compressor was not a good solution either, incredibly slow. then finally the postMessage sometimes crashes the entire code window. so even if improving log-parser / rewriting in wasm etc, doing so in the backend isn't really possible without changing the main scroller logic from vue-virtual-scroller to actual pagination. could also work but sounds pretty cumbersome