Fix search filter/jump radio buttons
They do not work at all and even if they are made working, they require the present container vscode-radio-group which does not seem to expose a way to enforce the layout to be horizontal even when the space is sparse, they always wrap below one another which must not happen. Also the padding between them is too big and this cannot be modified either. Omitting the group wrapper fails because you cannot set the name attribute on those vscode radio elements individually. So all in all these elements are not fit for use here.

Fixes #114